+++ integration_tests/test_gravity.py
@@ -870,44 +870,3 @@ def test_gravity_turn_bridge(gravity):
     # check no new batch is created
     rsp = cli.query_batches()
     assert len(rsp["batches"]) == 0
-
-
-def test_gravity_non_cosmos_denom(gravity):
-    if not gravity.cronos.enable_auto_deployment:
-        cronos_cli = gravity.cronos.cosmos_cli()
-
-        # Deploy a bad cosmos erc20 token
-        print("Deploy cosmos erc20 contract on ethereum")
-        tx_receipt = deploy_erc20(
-            gravity.contract, gravity.geth, "A", "A", "DOG", 6, KEYS["validator"]
-        )
-        assert tx_receipt.status == 1, "should success"
-
-        # Wait enough for orchestrator to relay the event
-        wait_for_new_blocks(cronos_cli, 30)
-
-        # deploy test erc20 contract and test that the orchestrator's still working
-        erc20 = deploy_contract(
-            gravity.geth,
-            CONTRACTS["TestERC20A"],
-        )
-
-        balance = erc20.caller.balanceOf(ADDRS["validator"])
-        assert balance == 100000000000000000000000000
-        amount = 1000
-
-        print("send to cronos crc20")
-        recipient = HexBytes(ADDRS["community"])
-        txreceipt = send_to_cosmos(
-            gravity.contract, erc20, gravity.geth, recipient, amount, KEYS["validator"]
-        )
-        assert txreceipt.status == 1, "should success"
-        assert erc20.caller.balanceOf(ADDRS["validator"]) == balance - amount
-
-        denom = f"gravity{erc20.address}"
-
-        def check_gravity_native_tokens():
-            "check the balance of gravity native token"
-            return cronos_cli.balance(eth_to_bech32(recipient), denom=denom) == amount
-
-        wait_for_fn("send-to-gravity-native", check_gravity_native_tokens)
